fix(columbia_merger): refactor common functions for harvard and columbia merger
update docstrings
update tests
move functions
