+++ cl/corpus_importer/tests.py
@@ -32,16 +32,10 @@ from cl.corpus_importer.management.commands.clean_up_mis_matched_dockets import 
     find_and_fix_mis_matched_dockets,
 )
 from cl.corpus_importer.management.commands.harvard_merge import (
-    ClusterSourceException,
-    DocketSourceException,
     combine_non_overlapping_data,
     fetch_non_harvard_data,
-    merge_case_names,
     merge_cluster_dates,
-    merge_docket_numbers,
-    merge_judges,
     merge_opinion_clusters,
-    merge_strings,
     update_cluster_source,
     update_docket_source,
 )
@@ -62,7 +56,15 @@ from cl.corpus_importer.management.commands.troller_bk import (
     merge_rss_data,
 )
 from cl.corpus_importer.tasks import generate_ia_json
-from cl.corpus_importer.utils import get_start_of_quarter
+from cl.corpus_importer.utils import (
+    ClusterSourceException,
+    DocketSourceException,
+    get_start_of_quarter,
+    merge_case_names,
+    merge_docket_numbers,
+    merge_judges,
+    merge_strings,
+)
 from cl.lib.pacer import process_docket_data
 from cl.lib.redis_utils import make_redis_interface
 from cl.lib.timezone_helpers import localize_date_and_time
@@ -2861,7 +2863,12 @@ class HarvardMergerTests(TestCase):
                 case_name_full=item[1].get("cl_case_name_full"),
             )
 
-            data_to_update = merge_case_names(cluster, item[0])
+            data_to_update = merge_case_names(
+                cluster,
+                item[0],
+                case_name_key="name_abbreviation",
+                case_name_full_key="name",
+            )
 
             self.assertEqual(
                 data_to_update.get("case_name", ""),
